new: dev: add doorstop TST document with test items, links, and references
* add test keywords and move some tests, update design doc
* add test doc to sphinx build
* note: the reference file test_extras.py causes a doorstop error
  but with the same tests and keyword moved to a new file it works
  just fine (using tests/test_timer.py as reference file)
