+++ docs/source/TUI_unit_tests.md
@@ -1,0 +1,24 @@
+# 1.0 TUI Automated Tests {#TST_001}
+
+TUI Automated Tests
+
+
+## 1.1 TST_002 {#TST_002}
+
+Test default settings:
+
+> `tests/test_helpers.py` (line 34)
+
+*Parent links: TUI_002*
+
+
+## 1.2 TST_003 {#TST_003}
+
+Test log tail handler:
+
+> `tests/test_helpers.py` (line 58)
+> `tests/test_timer.py` (line 15)
+
+*Parent links: TUI_005*
+
+
