+++ docs/swd/SDD_001.md
@@ -13,7 +13,7 @@ level: 1.0
 links: []
 normative: false
 ref: ''
-reviewed: vYvRCDeKKV70Bm4nZ9wxyTEO1-xxb9liiWTPvXnGziM=
+reviewed: 981rijaK3-MofaTSh7vTQ7kJkzkzXTXIAM_ce2b3zjo=
 ---
 
 # TUI Design Elements
@@ -72,13 +72,7 @@ sequenceDiagram
 
 #### SW dependencies
 
-```{figure} assets/tui_dependency_graph.svg
-:width: 90 %
-:align: center
-:alt: daemontui software units
-
-daemontui dependencies.
-```
+![primary software dependencies](assets/tui_dependency_graph.svg)
 
 <details>
   <summary>tui_dependency_graph source</summary>
