+++ tests/docs/TST_001.yml
@@ -1,0 +1,18 @@
+active: true
+derived: false
+doc:
+  by: sarnold
+  copyright: sarnold
+  major: '1'
+  minor: A
+  name: daemontui
+  ref: ''
+  title: TUI Integration and Unit Tests
+header: ''
+level: 1.0
+links: []
+normative: false
+ref: ''
+reviewed: vC-5jgxRWFbTDeXaPhB5sZmNMHZg3_upHdoP9hHyYvg=
+text: |
+  TUI Automated Tests
