+++ tox.ini
@@ -315,10 +315,10 @@ deps =
 
 commands_pre =
     docs: doorstop publish TUI docs/source/TUI_reqs_tree.md
-    #docs: doorstop publish LLT docs/source/TUI_unit_tests.md
+    docs: doorstop publish TST docs/source/TUI_unit_tests.md
     docs: doorstop publish SDD docs/source/TUI_sw_design.md
     docs: bash docs/sphinx_prep.sh docs/source/TUI_reqs_tree.md
-    #docs: bash docs/sphinx_prep.sh docs/source/TUI_unit_tests.md
+    docs: bash docs/sphinx_prep.sh docs/source/TUI_unit_tests.md
     docs: bash docs/sphinx_prep.sh docs/source/TUI_sw_design.md
 
 commands =
@@ -494,4 +494,5 @@ deps =
     pip>=21.1
 
 commands =
-    bash -c 'rm -rf build/ dist/ __pycache__/ src/*.egg-info/ docs/source/api/ .coverage*'
+    bash -c 'rm -rf src/*.egg-info dist/ build/ .coverage* coverage.xml'
+    bash -c 'rm -rf .pytest_cache __pycache__ docs/source/api docs/source/assets/*.svg'
